rag-chatbot/requirements.txt
streamlit>=1.32.0
langchain>=0.1.0
langchain-community>=0.0.20
sentence-transformers>=2.5.0
faiss-cpu>=1.7.4
PyMuPDF>=1.23.0
groq>=0.9.0
python-dotenv>=1.0.0
tiktoken>=0.6.0
numpy>=1.24.0